+++ src/main/java/digital/nogran/api/DigitalnogranApplication.java
@@ -1,4 +1,4 @@
-package br.com.digitalnogran;
+package digital.nogran.api;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
