+++ src/main/java/digital/nogran/api/model/Contato.java
@@ -1,4 +1,4 @@
-package br.com.digitalnogran.model;
+package digital.nogran.api.model;
 
 import java.time.LocalDateTime;
