+++ src/main/java/digital/nogran/api/model/EmailDetails.java
@@ -1,4 +1,4 @@
-package br.com.digitalnogran.model;
+package digital.nogran.api.model;
 
 //Importing required classes
