+++ src/main/java/digital/nogran/api/repository/ContatoRepository.java
@@ -1,12 +1,11 @@
-package br.com.digitalnogran.repository;
+package digital.nogran.api.repository;
 
 import java.util.List;
 
+import digital.nogran.api.model.Contato;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
-
-import br.com.digitalnogran.model.Contato;
 
 @Repository
 public interface ContatoRepository extends JpaRepository <Contato, Long> {
