+++ src/main/resources/application.properties
@@ -5,7 +5,7 @@ springdoc.api-docs.path=/v3/api-docs
 springdoc.swagger-ui.path=/swagger-ui.html
 springdoc.swagger-ui.operationsSorter=method
 springdoc.swagger-ui.disable-swagger-default-url=true
-springdoc.packagesToScan=br.com.digitalnogran.controller
+springdoc.packagesToScan=digital.nogran.api.controller
 
 #spring-boot-starter-mail properties
 spring.mail.host=smtp.gmail.com
