+++ src/test/java/digital/nogran/api/DigitalnogranApplicationTests.java
@@ -1,4 +1,4 @@
-package br.com.digitalnogran;
+package digital.nogran.api;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
